chore(addon): #22 move tests for content to content-test
