+++ content-test/lib/ActionManager.test.js
