+++ content-test/lib/ReduxChannel.test.js
