--- test/lib/.eslintignore
@@ -1,1 +1,0 @@
-./httpd.js
